reports/github-repo-stats/pyproject.toml
[tool.poetry]
name = "get-github-stats"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
python-dotenv = "^1.0.1"
rich = "^13.7.0"
tomli = "^2.0.1"
py-markdown-table = "^0.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
